+++ lib/apis/payment_api_manager.dart
@@ -41,6 +41,7 @@ class PaymentApiManager extends BaseApiManager {
       String merchantReference, Function success, Function fail) async {
     var result = await BaseApiManager.mainClient().query(QueryOptions(
         documentNode: GetOrderBillingStatusQuery().document,
+        fetchPolicy: FetchPolicy.noCache,
         variables:
             GetOrderBillingStatusArguments(merchantReference: merchantReference)
                 .toJson()));
@@ -49,6 +50,41 @@ class PaymentApiManager extends BaseApiManager {
     } else {
       success(GetOrderBillingStatus$Query.fromJson(result.data)
           .getOrderBillingStatus);
+    }
+  }
+
+  static Future<void> savedCardsApi(Function success, Function fail) async {
+    var result = await BaseApiManager.mainClient()
+        .query(QueryOptions(documentNode: SavedCardsQuery().document));
+    if (result.hasException) {
+      fail(ApiErrorHelper.handle(result));
+    } else {
+      success(
+          SavedCardList(SavedCards$Query.fromJson(result.data).getSavedCards));
+    }
+  }
+
+  static Future<void> placeSavedCreditCardOrderApi(
+      String addressId,
+      String voucherId,
+      String savedCardId,
+      String cvv,
+      Function success,
+      Function fail) async {
+    var result = await BaseApiManager.mainClient().mutate(MutationOptions(
+        documentNode: PlaceSavedCreditCardOrderMutation().document,
+        variables: PlaceSavedCreditCardOrderArguments(
+          addressId: addressId,
+          voucherId: voucherId,
+          savedCardId: savedCardId,
+          cardSecurityCode: cvv,
+        ).toJson()));
+    if (result.hasException) {
+      fail(ApiErrorHelper.handle(result));
+    } else {
+      success(PlaceSavedCreditCardOrderModel(
+          PlaceSavedCreditCardOrder$Mutation.fromJson(result.data)
+              .placeSavedCreditCardOrder));
     }
   }
 }